added checkId and HTTPStatus Code in Response

+++ src/main/java/com/game/controller/PlayerController.java
@@ -35,26 +35,34 @@ public class PlayerController {
     }*/
 
     @GetMapping("/players")
-    @ResponseStatus(HttpStatus.OK)
-    public List<Player> readAll(@RequestParam(value = "name", required = false) String name,
-                                @RequestParam(value = "title", required = false) String title,
-                                @RequestParam(value = "race", required = false) Race race,
-                                @RequestParam(value = "profession", required = false) Profession profession,
-                                @RequestParam(value = "after", required = false) Long after,
-                                @RequestParam(value = "before", required = false) Long before,
-                                @RequestParam(value = "banned", required = false) Boolean banned,
-                                @RequestParam(value = "minExperience", required = false) Integer minExperience,
-                                @RequestParam(value = "maxExperience", required = false) Integer maxExperience,
-                                @RequestParam(value = "minLevel", required = false) Integer minLevel,
-                                @RequestParam(value = "maxLevel", required = false) Integer maxLevel,
-                                @RequestParam(value = "order", required = false, defaultValue = "ID") PlayerOrder order,
-                                @RequestParam(value = "pageNumber", required = false, defaultValue = "0") Integer pageNumber,
-                                @RequestParam(value = "pageSize", required = false, defaultValue = "3") Integer pageSize
-    ) {
+//    @ResponseStatus(HttpStatus.OK)
+    public ResponseEntity<List<Player>> readAll(@RequestParam(value = "name", required = false) String name,
+                                                @RequestParam(value = "title", required = false) String title,
+                                                @RequestParam(value = "race", required = false) Race race,
+                                                @RequestParam(value = "profession", required = false) Profession profession,
+                                                @RequestParam(value = "after", required = false) Long after,
+                                                @RequestParam(value = "before", required = false) Long before,
+                                                @RequestParam(value = "banned", required = false) Boolean banned,
+                                                @RequestParam(value = "minExperience", required = false) Integer minExperience,
+                                                @RequestParam(value = "maxExperience", required = false) Integer maxExperience,
+                                                @RequestParam(value = "minLevel", required = false) Integer minLevel,
+                                                @RequestParam(value = "maxLevel", required = false) Integer maxLevel,
+                                                @RequestParam(value = "order", required = false, defaultValue = "ID") PlayerOrder order,
+                                                @RequestParam(value = "pageNumber", required = false, defaultValue = "0") Integer pageNumber,
+                                                @RequestParam(value = "pageSize", required = false, defaultValue = "3") Integer pageSize) {
 
         Pageable pageable = PageRequest.of(pageNumber, pageSize, Sort.by(order.getFieldName()));
 
-        return playerService.readAll(Specification.where(playerService.filterByName(name))
+        /*return playerService.readAll(Specification.where(playerService.filterByName(name))
+                .and(playerService.filterByTitle(title))
+                .and(playerService.filterByRace(race))
+                .and(playerService.filterByProfession(profession))
+                .and(playerService.filterByDate(after, before))
+                .and(playerService.filterByBanned(banned))
+                .and(playerService.filterByExperience(minExperience, maxExperience))
+                .and(playerService.filterByLevel(minLevel, maxLevel)), pageable).getContent();*/
+
+        List<Player> players = playerService.readAll(Specification.where(playerService.filterByName(name))
                 .and(playerService.filterByTitle(title))
                 .and(playerService.filterByRace(race))
                 .and(playerService.filterByProfession(profession))
@@ -62,8 +70,9 @@ public class PlayerController {
                 .and(playerService.filterByBanned(banned))
                 .and(playerService.filterByExperience(minExperience, maxExperience))
                 .and(playerService.filterByLevel(minLevel, maxLevel)), pageable).getContent();
-    }
 
+        return new ResponseEntity<>(players, HttpStatus.OK);
+    }
 
     @GetMapping("/players/count")
     @ResponseStatus(HttpStatus.OK)
@@ -89,20 +98,51 @@ public class PlayerController {
                 .and(playerService.filterByLevel(minLevel, maxLevel))).size();
     }
 
+    @GetMapping("players/{id}")
+    public ResponseEntity<Player> read(@PathVariable(value = "id") String idString) {
+        Long id = playerService.checkId(idString);
+
+        if (id == -1L)
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+
+        Player player = playerService.read(id);
+
+        return player != null ? new ResponseEntity<>(player, HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
     @PostMapping("/players")
-    @ResponseStatus(HttpStatus.OK)
-    @ResponseBody
-    public Player addShip(@RequestBody Player player) {
+    public ResponseEntity<Player> addShip(@RequestBody Player player) {
 
-        return playerService.create(player);
+        Player createdPlayer = playerService.create(player);
 
+        return createdPlayer != null ? new ResponseEntity<>(createdPlayer, HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+    }
+
+    @PutMapping("players/{id}")
+    public ResponseEntity<Player> update(@RequestBody Player player, @PathVariable(value = "id") String idString) {
+        Long id = playerService.checkId(idString);
+
+        if (id == -1L)
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+
+        Player updatedPlayer = playerService.update(player, id);
+
+        return player != null ? new ResponseEntity<>(updatedPlayer, HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("players/{id}")
-    public ResponseEntity<?> delete(@PathVariable(value = "id") Long id) {
+    public ResponseEntity<?> delete(@PathVariable(value = "id") String idString) {
+        Long id = playerService.checkId(idString);
+
+        if (id == -1L)
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+
         Player player = playerService.delete(id);
 
-        return player == null ? new ResponseEntity<>(HttpStatus.NOT_FOUND) : new ResponseEntity<>(HttpStatus.OK);
+        return player != null ? new ResponseEntity<>(HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
-
 }
